Add resources page, add socat port listener script, clean up site
Repurpose the "News" page into the "Resources" page for competitors to reference during the CTF.
Replace the "CTF Placeholder" with "Mini CTF". Clean up the site - recolor the navbar.

Manually add port to iframe on the shell page. There is an issue where the web shell is hosted
at a different port than 80. The shell page assumes that the web shell is at port 80. If the
port is added to the end of the host, then the shell management is unable to connect to ssh
into the shell box (since then it tries connecting to HOST:WEB_PORT:SSH_PORT).

Add socat listener script for use in making the vagrant machine services accessible externally.